Renamed Tipi\\ namespace and classes to Totem\\
