+++ composer.json
@@ -5,7 +5,7 @@
     "description": "PHP Module for Totem",
     "autoload": {
         "psr-4": {
-            "Tipi\\": "lib/"
+            "Totem\\": "lib/"
         }
     },
     "require": {
